requirements.txt
# Keep crewai at version 0.130.0
# Other package versions are flexible and can be changed
# Only change crewai version if there are critical dependency conflicts that cannot be resolved by other means
crewai==0.130.0
crewai-tools==0.47.1
fastapi
uvicorn
python-dotenv
pypdf
python-multipart
langchain-google-genai
redis
celery
sqlmodel
sqlalchemy